implement arithmetic tests for period dtype
